Triangles should accept non-integer side lengths.

+++ src/main/java/eu/throup/triangles/Triangle.java
@@ -1,11 +1,11 @@
 package eu.throup.triangles;
 
 public class Triangle {
-    private final int side1;
-    private final int side2;
-    private final int side3;
+    private final double side1;
+    private final double side2;
+    private final double side3;
 
-    public Triangle(int side1, int side2, int side3) {
+    public Triangle(double side1, double side2, double side3) {
         this.side1 = side1;
         this.side2 = side2;
         this.side3 = side3;
